fix: #2140 1BP v4 discovery + OnebpArch remap; #2141 census tokens jugnuvr/m2r; #2143 overflow/format fixes (#2162)
* fix(engine): #2140 1BP v4 discovery + OnebpArch remap; #2141 census tokens jugnuvr/m2r

- model_discovery: accept ONEBP_VERSION (v4 dedup-alias index) — the
  version>3 cap hid every current v4 conversion from discovery
- replace the stale arch-enum switch (1->llama) with the real OnebpArch
  enum; ONEBP_MOE now probes the tensor index (GDN signature -> qwen35)
  then name/dims inference (zaya/cohere2moe/qwen3moe/phi_moe), and
  refuses loudly when unclassifiable instead of silently misrouting
- bitnet_model: registry tokens RCPP_ARCH_JUGNUVR / RCPP_ARCH_M2R with
  arch-string + stripped-name mappings (census coverage gate)

* fix(sec): #2143 triage — widen overflow-prone int mults, printf format fixes

- baretorch_engine: U/R/yloc sized by H*r_/H*dh computed in int before
  vector size_type conversion (integer-multiplication-cast-to-long)
- npu_engine_universal: two scan loops bounded by
  (uint32_t)MD*(uint32_t)bC_nd — widen to size_t before multiplying
- diag_fused_weight_accuracy: %lld with int 'sat' (wrong-type-format,
  UB) and %ld with long long 'cnt' — corrected

---------

+++ src/baretorch_engine.cpp
@@ -301,7 +301,7 @@ private:
     // ── cs_lrad layer: exact chunked math, incremental ──
     void cs_lrad_step(BrLayer& ly, const float* xn, float* out) {
         // project this token
-        std::vector<float> q(D), k(D), v(D), U(H * r_), R(H * r_), gate(H), beta(H), lg(H), sw(D);
+        std::vector<float> q(D), k(D), v(D), U((size_t)H * r_), R((size_t)H * r_), gate(H), beta(H), lg(H), sw(D);
         mm(ly.W_q, xn, D, D, q.data());
         mm(ly.W_k, xn, D, D, k.data());
         mm(ly.W_v, xn, D, D, v.data());
@@ -345,7 +345,7 @@ private:
 
         // Y_local: chunked decay-link causal attention over current chunk
         //   out[h,d] = sum_{j<=pos} exp(min(lam_pos(h)-lam_j(h),0)) * (q_pos . k_j) * v_j / sqrt(dh)
-        std::vector<float> yloc(H * dh, 0.0f);
+        std::vector<float> yloc((size_t)H * dh, 0.0f);
         const float* qp = ly.cq[pos].data();
         for (int j = 0; j <= pos; j++) {
             const float* kj = ly.ck[j].data();
